fix(common-css): add headings to common CSS [ci visual]
