+++ packages/common-css/src/_common-variables.scss
@@ -59,6 +59,13 @@ $sap-text-monospaced-bold: 'monospaced-bold';
 $sap-text-black: 'black';
 $sap-text-header: 'header';
 
+$sap-heading-h6: 'h6';
+$sap-heading-h5: 'h5';
+$sap-heading-h4: 'h4';
+$sap-heading-h3: 'h3';
+$sap-heading-h2: 'h2';
+$sap-heading-h1: 'h1';
+
 // Border-radius
 $sap-corner-border-radius: (
   "element": var(--sapElement_BorderCornerRadius),
