--- packages/common-css/src/sap-title.scss
@@ -1,45 +1,0 @@
-@import 'common-settings';
-@import 'common-mixins';
-
-$block: #{$sap-namespace}-title;
-
-.#{$block} {
-  @include sap-reset();
-
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  font-size: 1rem;
-  color: var(--sapTextColor);
-  font-family: var(--sapFontHeaderFamily);
-
-  &--h1 {
-    font-size: var(--sapFontHeader1Size);
-  }
-
-  &--h2 {
-    font-size: var(--sapFontHeader2Size);
-  }
-
-  &--h3 {
-    font-size: var(--sapFontHeader3Size);
-  }
-
-  &--h4 {
-    font-size: var(--sapFontHeader4Size);
-  }
-
-  &--h5 {
-    font-size: var(--sapFontHeader5Size);
-  }
-
-  &--h6 {
-    font-size: var(--sapFontHeader6Size);
-  }
-
-  &--wrap {
-    overflow: auto;
-    white-space: normal;
-    text-overflow: initial;
-  }
-}
