+++ packages/common-css/stories/heading/levels.example.html
@@ -1,0 +1,6 @@
+<h1 class="sap-heading-1">Heading 1</h1>
+<h2 class="sap-heading-2">Heading 2</h2>
+<h3 class="sap-heading-3">Heading 3</h3>
+<h4 class="sap-heading-4">Heading 4</h4>
+<h5 class="sap-heading-5">Heading 5</h5>
+<h6 class="sap-heading-6">Heading 6</h6>
