--- packages/common-css/stories/title/elision.example.html
@@ -1,8 +1,0 @@
-<div style="width: 300px">
-    <h1 class="sap-title sap-title--h1">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h1>
-    <h2 class="sap-title sap-title--h2">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h2>
-    <h3 class="sap-title sap-title--h3">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h3>
-    <h4 class="sap-title sap-title--h4">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h4>
-    <h5 class="sap-title sap-title--h5">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h5>
-    <h6 class="sap-title sap-title--h6">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h6>
-</div>
