--- packages/common-css/stories/title/levels.example.html
@@ -1,6 +1,0 @@
-<h1 class="sap-title sap-title--h1">Title Level 1</h1>
-<h2 class="sap-title sap-title--h2">Title Level 2</h2>
-<h3 class="sap-title sap-title--h3">Title Level 3</h3>
-<h4 class="sap-title sap-title--h4">Title Level 4</h4>
-<h5 class="sap-title sap-title--h5">Title Level 5</h5>
-<h6 class="sap-title sap-title--h6">Title Level 6</h6>
