--- packages/common-css/stories/title/wrapping.example.html
@@ -1,8 +1,0 @@
-<div style="width: 300px">
-    <h1 class="sap-title sap-title--h1 sap-title--wrap">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h1>
-    <h2 class="sap-title sap-title--h2 sap-title--wrap">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h2>
-    <h3 class="sap-title sap-title--h3 sap-title--wrap">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h3>
-    <h4 class="sap-title sap-title--h4 sap-title--wrap">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h4>
-    <h5 class="sap-title sap-title--h5 sap-title--wrap">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h5>
-    <h6 class="sap-title sap-title--h6 sap-title--wrap">Lorem ipsum dolor sit amet, consectetur adipiscing elit</h6>
-</div>
